Refactor to MVC structure and split frontend components
