+++ client/src/types/Game.ts
@@ -1,0 +1,10 @@
+export interface Game {
+  id: string
+  name: string
+  yearpublished?: string
+  image?: string
+  minage?: string
+  minplayers?: string
+  maxplayers?: string
+  categories?: string[]
+}
