+++ server/routes/searchRoutes.js
@@ -1,0 +1,8 @@
+import { Router } from 'express'
+import { searchGames } from '../controllers/searchController.js'
+
+const router = Router()
+
+router.get('/', searchGames)
+
+export default router
